# 02/openapi.yaml
openapi: "3.0.3"
info:
  title: Kanvas API
  description: REST API for the Kanvas Kanban board application
  version: "1.0.0"
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /api/boards/{boardId}/cards:
    post:
      summary: Create a card on a board
      tags: [Cards]
      parameters:
        - $ref: "#/components/parameters/BoardId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [column, title]
              properties:
                column:
                  $ref: "#/components/schemas/Column"
                title:
                  type: string
                  maxLength: 200
                description:
                  type: string
                due_date:
                  type: string
                  format: date
                  nullable: true
                assignee:
                  type: string
      responses:
        "201":
          description: Card created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/boards/{boardId}/reorder:
    patch:
      summary: Reorder cards within a column
      tags: [Cards]
      parameters:
        - $ref: "#/components/parameters/BoardId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [column, card_ids]
              properties:
                column:
                  $ref: "#/components/schemas/Column"
                card_ids:
                  type: array
                  items:
                    type: integer
      responses:
        "200":
          description: Reorder successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/cards/{cardId}:
    put:
      summary: Update a card
      tags: [Cards]
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 200
                description:
                  type: string
                due_date:
                  type: string
                  format: date
                  nullable: true
                assignee:
                  type: string
      responses:
        "200":
          description: Card updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a card
      tags: [Cards]
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Card deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/cards/{cardId}/move:
    patch:
      summary: Move a card to a different column
      tags: [Cards]
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_column, new_position]
              properties:
                new_column:
                  $ref: "#/components/schemas/Column"
                new_position:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: Card moved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  schemas:
...
    Card:
      type: object
      properties:
        id:
          type: integer
        board_id:
          type: integer
        column:
          $ref: "#/components/schemas/Column"
        title:
          type: string
        description:
          type: string
        due_date:
          type: string
          format: date
          nullable: true
        assignee:
          type: string
        position:
          type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Column:
      type: string
      enum: [todo, in_progress, done]
  parameters:
    BoardId:
      name: boardId
      in: path
      required: true
      schema:
        type: integer
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
...